Address Lukas' comment about comparison scripts
