+++ gitlab-cpu-pipeline.yml
@@ -141,7 +141,7 @@ check_faultoutput:
         - ls
         - cd output-tpv{$tpv}-${precision}
         - ls
-        - python3 ../postprocessing/validation/compare-faults/compare.py ./${fault_file} ./precomputed/${fault_file} --epsilon ${epsilon}
+        - python3 ../postprocessing/validation/compare-faults.py ./${fault_file} ./precomputed/${fault_file} --epsilon ${epsilon}
 
 
 check_receivers:
@@ -161,7 +161,7 @@ check_receivers:
         - cd output-tpv{$tpv}-${precision}
         - ls
         - if [ ${tpv} = 5 ] || [ ${tpv} = 5-nuc ] || [ ${tpv} = 6 ] || [ ${tpv} = 16 ]; then mode=lsw; elif [ ${tpv} = 105 ]; then mode=tp; else mode=rs; fi;
-          python3 ../postprocessing/validation/compare-receivers/compare.py . ./precomputed --epsilon ${epsilon} --mode $mode
+          python3 ../postprocessing/validation/compare-receivers.py . ./precomputed --epsilon ${epsilon} --mode $mode
 
 
 check_energies:
@@ -181,4 +181,4 @@ check_energies:
         - ls
         - cd output-tpv{$tpv}-${precision}
         - ls
-        - python3 ../postprocessing/validation/compare-energies/compare.py ./${energy_file} ./precomputed/${energy_file} --epsilon ${epsilon}
+        - python3 ../postprocessing/validation/compare-energies.py ./${energy_file} ./precomputed/${energy_file} --epsilon ${epsilon}
